Time: 7 ms (87.43%), Space: 17.3 MB (92.36%) - LeetHub

+++ 0662-maximum-width-of-binary-tree/0662-maximum-width-of-binary-tree.cpp
@@ -13,7 +13,7 @@ class Solution {
 public:
     int widthOfBinaryTree(TreeNode* root) {
         if(!root) return 0;
-        queue<pair<TreeNode*,long>> q;
+        queue<pair<TreeNode*,int>> q;
         q.push({root,0});
         int mlen = 1;
         while(!q.empty()){
